Core functionality of VCC completed

+++ NeoV14/events/voiceStateUpdate.js
@@ -1,6 +1,6 @@
 const { vcCreator } = require("../libs/gdb");
 const { ChannelType } = require('discord.js');
-const TIMEOUT_TIME = 15000; //300000 = 5min
+const TIMEOUT_TIME = 5000; //300000 = 5min
 
 module.exports = {
 	name: 'voiceStateUpdate',
@@ -8,41 +8,37 @@ module.exports = {
 	async run (oldState, newState, token) {
 		let gvc = await vcCreator.findByPk(newState.guild.id);
 		if(!gvc) return;
-		//console.log(gvc.spawnedVCs.filter(vc => vc.id == newState.channelId))
 
 		if(newState.channelId == oldState.channelId) return; //means user deafened or muted
-		if(newState.channelId != oldState.channelId) { //user left the channel, check if its part of the vccreator system
-			let vcObject = gvc.spawnedVCs.filter(vc => vc.id == oldState.channelId);
+	
+		if(newState.channelId != oldState.channelId && oldState.channelId != null) { //user left the channel, check if its part of the vccreator system
+			let vcObject = gvc.spawnedVCs.filter(vc => vc == oldState.channelId);
 			if(vcObject[0] != undefined){ // I love filters <3
-				console.log('User left a spawned vc channel');
+				//If the channel is vc
 				let channel = oldState.guild.channels.cache.get(oldState.channelId);
 
 				if(channel.members.size == 0){ // If no one is in the channel, delete it after 5 minutes of inactivity
-					vcObject.timeoutid = setTimeout(() => { // this timeout is checked and removed in next elseif block when fired
-						if(channel.members.size == 0) {
+					
+					setTimeout(async () => { // this timeout is checked and removed in next elseif block when fired
+						let gvc = await vcCreator.findByPk(newState.guild.id);
+						channel = oldState.guild.channels.cache.get(channel.id);
+						if(channel && channel.members.size == 0) {
 							try { channel.delete(); } catch {};
-							let temp = gvc.spawnedVCs.filter(vc => vc.id != oldState.channelId);
-							if(temp.length == 0) gvc.spawnedVCs = [];
-							else gvc.spawnedVCs = [...temp];
-							gvc.save();
+							let temp = gvc.spawnedVCs.filter(vc => vc != oldState.channelId);
+							gvc.spawnedVCs = [...temp];
+							// if(temp.length == 0) gvc.spawnedVCs = [];
+							// else gvc.spawnedVCs = [...temp];
+
+							await gvc.save();
 						}
 					}, TIMEOUT_TIME);
 
+					let temp = gvc.spawnedVCs.filter(vc => vc != oldState.channelId);
+					gvc.spawnedVCs = [...temp, vcObject[0]];
 					await gvc.save();
 				}
 			}
-		//check if the user joined a spawned vc channel
-		} else if(JSON.stringify(gvc.spawnedVCs.filter(vc => vc.id == newState.channelId)) != '[]'){
-			console.log('User joined a spawned vc channel')
-			let vcObject = gvc.spawnedVCs.filter(vc => vc.id == newState.channelId); 
-			if(vcObject.timeoutid != null) { //if the timeout is set, clear it and remove it from the object
-				clearTimeout(vcObject.timeoutid); 
-				vcObject.timeoutid = null;
-				let temp = gvc.spawnedVCs.filter(vc => vc.id != newState.channelId);
-				gvc.spawnedVCs = [...temp, vcObject];
-				await gvc.save();
-			}
-		}
+		} 
 
 		//if the user joined the vccreator channel, create a new vc for them
 		if(newState.channelId == gvc.channel){
@@ -55,10 +51,10 @@ module.exports = {
 			let vc = await newState.guild.channels.create({name: 'testy', type: ChannelType.GuildVoice, parent: gvc.category});
 			// TODO: figure out how to uniquely identify the name of the channel. Remove it from being called testy. maybe use users name?
 			// Then figure out logic for users changing the name. Maybe a DM or a ping in the VC chat
-			gvc.spawnedVCs = [...gvc.spawnedVCs, {id: vc.id, timeoutid: null}];
+			gvc.spawnedVCs = [...gvc.spawnedVCs, vc.id];
 			await gvc.save();
 
-			newState.member.voice.setChannel(vc);
+			try { newState.member.voice.setChannel(vc) } catch {};
 		}
 	}
 }
